+++ task.go
@@ -5,7 +5,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/gtkit/json"
+	"github.com/gtkit/json/v2"
 	"github.com/hibiken/asynq"
 )